+++ src/librustc_target/abi/call/aarch64.rs
@@ -1,4 +1,4 @@
-// Copyright 2012-2013 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2015 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,10 +8,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use abi::{FnType, ArgType, LayoutExt, Reg, RegKind, Uniform};
-use rustc_target::abi::{HasDataLayout, LayoutOf, TyLayout, TyLayoutMethods};
-use rustc_target::spec::HasTargetSpec;
-use llvm::CallConv;
+use abi::call::{FnType, ArgType, Reg, RegKind, Uniform};
+use abi::{HasDataLayout, LayoutOf, TyLayout, TyLayoutMethods};
 
 fn is_homogeneous_aggregate<'a, Ty, C>(cx: C, arg: &mut ArgType<'a, Ty>)
                                      -> Option<Uniform> 
@@ -43,7 +41,7 @@ fn is_homogeneous_aggregate<'a, Ty, C>(cx: C, arg: &mut ArgType<'a, Ty>)
     })
 }
 
-fn classify_ret_ty<'a, Ty, C>(cx: C, ret: &mut ArgType<'a, Ty>, vfp: bool)
+fn classify_ret_ty<'a, Ty, C>(cx: C, ret: &mut ArgType<'a, Ty>) 
     where Ty: TyLayoutMethods<'a, C> + Copy,
           C: LayoutOf<Ty = Ty, TyLayout = TyLayout<'a, Ty>> + HasDataLayout
 {
@@ -51,24 +49,23 @@ fn classify_ret_ty<'a, Ty, C>(cx: C, ret: &mut ArgType<'a, Ty>, vfp: bool)
         ret.extend_integer_width_to(32);
         return;
     }
-
-    if vfp {
-        if let Some(uniform) = is_homogeneous_aggregate(cx, ret) {
-            ret.cast_to(uniform);
-            return;
-        }
+    if let Some(uniform) = is_homogeneous_aggregate(cx, ret) {
+        ret.cast_to(uniform);
+        return;
     }
-
     let size = ret.layout.size;
     let bits = size.bits();
-    if bits <= 32 {
+    if bits <= 128 {
         let unit = if bits <= 8 {
             Reg::i8()
         } else if bits <= 16 {
             Reg::i16()
-        } else {
+        } else if bits <= 32 {
             Reg::i32()
+        } else {
+            Reg::i64()
         };
+
         ret.cast_to(Uniform {
             unit,
             total: size
@@ -78,7 +75,7 @@ fn classify_ret_ty<'a, Ty, C>(cx: C, ret: &mut ArgType<'a, Ty>, vfp: bool)
     ret.make_indirect();
 }
 
-fn classify_arg_ty<'a, Ty, C>(cx: C, arg: &mut ArgType<'a, Ty>, vfp: bool) 
+fn classify_arg_ty<'a, Ty, C>(cx: C, arg: &mut ArgType<'a, Ty>) 
     where Ty: TyLayoutMethods<'a, C> + Copy,
           C: LayoutOf<Ty = Ty, TyLayout = TyLayout<'a, Ty>> + HasDataLayout
 {
@@ -86,38 +83,42 @@ fn classify_arg_ty<'a, Ty, C>(cx: C, arg: &mut ArgType<'a, Ty>, vfp: bool)
         arg.extend_integer_width_to(32);
         return;
     }
-
-    if vfp {
-        if let Some(uniform) = is_homogeneous_aggregate(cx, arg) {
-            arg.cast_to(uniform);
-            return;
-        }
+    if let Some(uniform) = is_homogeneous_aggregate(cx, arg) {
+        arg.cast_to(uniform);
+        return;
     }
+    let size = arg.layout.size;
+    let bits = size.bits();
+    if bits <= 128 {
+        let unit = if bits <= 8 {
+            Reg::i8()
+        } else if bits <= 16 {
+            Reg::i16()
+        } else if bits <= 32 {
+            Reg::i32()
+        } else {
+            Reg::i64()
+        };
 
-    let align = arg.layout.align.abi();
-    let total = arg.layout.size;
-    arg.cast_to(Uniform {
-        unit: if align <= 4 { Reg::i32() } else { Reg::i64() },
-        total
-    });
+        arg.cast_to(Uniform {
+            unit,
+            total: size
+        });
+        return;
+    }
+    arg.make_indirect();
 }
 
 pub fn compute_abi_info<'a, Ty, C>(cx: C, fty: &mut FnType<'a, Ty>) 
     where Ty: TyLayoutMethods<'a, C> + Copy,
-          C: LayoutOf<Ty = Ty, TyLayout = TyLayout<'a, Ty>> + HasDataLayout + HasTargetSpec
+          C: LayoutOf<Ty = Ty, TyLayout = TyLayout<'a, Ty>> + HasDataLayout
 {
-    // If this is a target with a hard-float ABI, and the function is not explicitly
-    // `extern "aapcs"`, then we must use the VFP registers for homogeneous aggregates.
-    let vfp = cx.target_spec().llvm_target.ends_with("hf")
-        && fty.cconv != CallConv::ArmAapcsCallConv
-        && !fty.variadic;
-
     if !fty.ret.is_ignore() {
-        classify_ret_ty(cx, &mut fty.ret, vfp);
+        classify_ret_ty(cx, &mut fty.ret);
     }
 
     for arg in &mut fty.args {
         if arg.is_ignore() { continue; }
-        classify_arg_ty(cx, arg, vfp);
+        classify_arg_ty(cx, arg);
     }
 }
