+++ src/librustc_target/abi/call/hexagon.rs
@@ -1,4 +1,4 @@
-// Copyright 2018 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2012-2013 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,14 +8,24 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use abi::{FnType, ArgType};
+#![allow(non_upper_case_globals)]
+
+use abi::call::{FnType, ArgType};
 
 fn classify_ret_ty<Ty>(ret: &mut ArgType<Ty>) {
-    ret.extend_integer_width_to(32);
+    if ret.layout.is_aggregate() && ret.layout.size.bits() > 64 {
+        ret.make_indirect();
+    } else {
+        ret.extend_integer_width_to(32);
+    }
 }
 
 fn classify_arg_ty<Ty>(arg: &mut ArgType<Ty>) {
-    arg.extend_integer_width_to(32);
+    if arg.layout.is_aggregate() && arg.layout.size.bits() > 64 {
+        arg.make_indirect();
+    } else {
+        arg.extend_integer_width_to(32);
+    }
 }
 
 pub fn compute_abi_info<Ty>(fty: &mut FnType<Ty>) {
@@ -24,7 +34,9 @@ pub fn compute_abi_info<Ty>(fty: &mut FnType<Ty>) {
     }
 
     for arg in &mut fty.args {
-        if arg.is_ignore() { continue; }
+        if arg.is_ignore() {
+            continue;
+        }
         classify_arg_ty(arg);
     }
 }
