+++ src/librustc_target/abi/call/msp430.rs
@@ -1,4 +1,4 @@
-// Copyright 2016 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2012-2013 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,24 +8,30 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// Reference: PTX Writer's Guide to Interoperability
-// http://docs.nvidia.com/cuda/ptx-writers-guide-to-interoperability
+// Reference: MSP430 Embedded Application Binary Interface
+// http://www.ti.com/lit/an/slaa534/slaa534.pdf
 
-use abi::{ArgType, FnType, LayoutExt};
+use abi::call::{ArgType, FnType};
 
+// 3.5 Structures or Unions Passed and Returned by Reference
+//
+// "Structures (including classes) and unions larger than 32 bits are passed and
+// returned by reference. To pass a structure or union by reference, the caller
+// places its address in the appropriate location: either in a register or on
+// the stack, according to its position in the argument list. (..)"
 fn classify_ret_ty<Ty>(ret: &mut ArgType<Ty>) {
-    if ret.layout.is_aggregate() && ret.layout.size.bits() > 64 {
+    if ret.layout.is_aggregate() && ret.layout.size.bits() > 32 {
         ret.make_indirect();
     } else {
-        ret.extend_integer_width_to(64);
+        ret.extend_integer_width_to(16);
     }
 }
 
 fn classify_arg_ty<Ty>(arg: &mut ArgType<Ty>) {
-    if arg.layout.is_aggregate() && arg.layout.size.bits() > 64 {
+    if arg.layout.is_aggregate() && arg.layout.size.bits() > 32 {
         arg.make_indirect();
     } else {
-        arg.extend_integer_width_to(64);
+        arg.extend_integer_width_to(16);
     }
 }
